Revert "Fix type mismatch in mgen_data::hostile_at()."
This reverts commit 4f45919f3dedc18a2aed6a22d152ceac6d1f6b85.

There's actually no mismatch (oops); monflags holds MG_-type flags, so
it's only 32-bit.

+++ crawl-ref/source/mgen_data.h
@@ -170,7 +170,7 @@ struct mgen_data
                                 int abj = 0,
                                 int st = 0,
                                 const coord_def &p = coord_def(-1, -1),
-                                uint64_t monflags = 0,
+                                unsigned monflags = 0,
                                 god_type ngod = GOD_NO_GOD,
                                 monster_type base = MONS_NO_MONSTER)
